requirements.txt
# Web scraping
playwright>=1.41.0
html5lib>=1.1

# Search engine
duckduckgo-search>=7.2.1

# LLM integration
openai>=1.59.8 # o1 support
anthropic>=0.42.0
python-dotenv>=1.0.0

# Testing
unittest2>=1.1.0
pytest>=8.0.0
pytest-asyncio>=0.23.5

# Google Generative AI
google-generativeai

# gRPC, for Google Generative AI preventing WARNING: All log messages before absl::InitializeLog() is called are written to STDERR
grpcio==1.70.0

# Dependências principais
crewai>=0.11.0
openai>=1.65.4
python-dotenv>=1.0.1
python-wordpress-xmlrpc>=2.3

# Utilitários
requests>=2.31.0
beautifulsoup4>=4.12.0
python-slugify>=8.0.0

# Ferramentas de desenvolvimento
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
isort>=5.12.0
pytest>=7.4.0
pytest-cov>=4.1.0

# Documentação
mkdocs>=1.5.3
mkdocs-material>=9.5.3

# New dependencies
langchain>=0.1.0
Pillow>=11.1.0
